AEGIS/requirements.txt
# AEGIS V4 — production runtime

# Core API
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-dotenv>=1.0
httpx>=0.27
sse-starlette>=2.0

# LLM client (OpenAI-compatible — works with LM Studio and Ollama)
openai>=1.30

# Embeddings (local, in-process, no LLM dependency)
sentence-transformers>=2.7
# nomic-embed-text-v1.5 trust_remote_code path requires einops
einops>=0.8

# Persistence
# pysqlcipher3>=1.2.1     # optional — falls back to plain sqlite3 if absent
#                          # (records.py logs a warning so this is never silent)

# RAG
chromadb>=0.5

# PDF ingestion
pypdf>=4.2
pdfplumber>=0.11

# Speech-to-text
faster-whisper>=1.0
numpy>=1.26
imageio-ffmpeg>=0.5

# Telemetry
psutil>=5.9

# V5 — uncomment when activating live inference
# (V4 ships with INFERENCE_MODE=mock; no LLM dependency)
# ollama>=0.3            # only if using the official Python client
# cryptography>=42.0     # for V5's Ed25519 signing chain upgrade

# Tests
pytest>=8.0
